Move convert_bool to utils.py
