+++ egress_backend/egress_backend_stack.py
@@ -40,24 +40,7 @@ from egress_backend.components.email_identity.email_identity_verification_cr imp
     EmailIdentityVerificationCustomResource,
 )
 
-
-def convert_bool(s, rtype):
-    """
-    Returns:
-        rtype=str: 'true' or 'false'
-        type=bool: true or false
-    """
-    if rtype not in (str, bool):
-        raise ValueError(f"Invalid return type: {rtype}")
-    if (isinstance(s, str) and s.lower() == "true") or s is True:
-        if rtype == str:
-            return "true"
-        return True
-    if (isinstance(s, str) and s.lower() == "false") or s is False:
-        if rtype == str:
-            return "false"
-        return False
-    raise ValueError(f"Invalid boolean string: {s}")
+from .components.utils import convert_bool
 
 
 class EgressBackendStack(Stack):
@@ -1221,7 +1204,9 @@ class EgressBackendStack(Stack):
         # add additional Idp as per configuration
         supported_idps = ["COGNITO"]
 
-        if self.node.try_get_context(env_id).get("custom_idp").get("is_enabled"):
+        if convert_bool(
+            self.node.try_get_context(env_id).get("custom_idp").get("is_enabled"), bool
+        ):
             cognito.CfnUserPoolIdentityProvider(
                 self,
                 "CustomIdentityProvider",
